requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.4
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.9.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.4
pylint==2.4.4
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
